amd64 (and i386): all pointers stored in java now stored as jlongs, use new mvutil data field.
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@15921 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ jvector/impl/com/untangle/jvector/OutgoingSocketQueue.java
@@ -315,6 +315,6 @@ public class OutgoingSocketQueue extends Source
         }
     }
 
-    private native int create();
-    private native void mvpollNotifyObservers( int pointer, int eventMask );
+    private native long create();
+    private native void mvpollNotifyObservers( long pointer, int eventMask );
 }
